+++ amber/src/test/java/com/hartwig/hmftools/amber/BaseDepthIntersectFilterTest.java
@@ -76,7 +76,7 @@ public class BaseDepthIntersectFilterTest
     {
         return ModifiableBaseDepth.create()
                 .setChromosome(chromosome)
-                .setPosition(random.nextLong())
+                .setPosition(random.nextInt())
                 .setRef(BaseDepth.Base.A)
                 .setAlt(BaseDepth.Base.T)
                 .setReadDepth(random.nextInt())
